+++ src/scrapers/movie/imdb/ImdbMovie.cpp
@@ -184,11 +184,11 @@ QVector<ScraperSearchResult> ImdbMovie::parseSearch(const QString& html)
 
     if (html.contains("Including Adult Titles")) {
         // Search result table from "https://www.imdb.com/search/title/?title=..."
-        rx.setPattern(R"(<a href="/title/(tt[\d]+)/[^"]*"\n>([^<]*)</a>\n.*(?: \(I+\) |>)\(([0-9]*).*\))");
+        rx.setPattern(R"(<a href="/title/(tt[\d]+)/[^"]*"\n>([^<]*)</a>\n\s*<span[^>]*>\((\d+)\)</span>)");
     } else {
         // Search result table from "https://www.imdb.com/find?q=..."
-        rx.setPattern("<td class=\"result_text\"> <a href=\"/title/([t]*[\\d]+)/[^\"]*\" >([^<]*)</a>(?: \\(I+\\) | "
-                      ")\\(([0-9]*)\\) (?:</td>|<br/>)");
+        rx.setPattern("<td class=\"result_text\"> <a href=\"/title/([t]*[\\d]+)/[^\"]*\" >([^<]*)</a>(?: \\(I+\\)"
+                      ")? \\(([0-9]+)\\) (?:</td>|<br/>)");
     }
 
     rx.setPatternOptions(QRegularExpression::DotMatchesEverythingOption | QRegularExpression::InvertedGreedinessOption);